+++ eyes.appium.java/src/test/java/com/applitools/eyes/appium/android/AndroidTestSetup.java
@@ -25,7 +25,7 @@ public abstract class AndroidTestSetup extends TestSetup {
 
     @Override
     protected void setAppCapability() {
-        // To run locally use https://applitools.bintray.com/Examples/android/1.2/app_android.apk
+        // To run locally use https://applitools.jfrog.io/artifactory/Examples/android/1.2/app_android.apk
         capabilities.setCapability("app", "app_android");
     }