+++ eyes.appium.java/src/test/java/com/applitools/eyes/appium/android/ScrollRootElementTest.java
@@ -23,7 +23,7 @@ public class ScrollRootElementTest extends AndroidTestSetup {
 
     @Override
     protected void setAppCapability() {
-        // To run locally use https://applitools.bintray.com/Examples/androidx/1.0.0/app_androidx.apk
+        // To run locally use https://applitools.jfrog.io/artifactory/Examples/androidx/1.0.0/app_androidx.apk
         capabilities.setCapability("app", "app_androidx");
     }